strategies tests moved to a directory
